(feature_rpc_settings): move RpcButtons to domain module
