+++ feature_rpc_base/build.gradle.kts
@@ -35,7 +35,6 @@ dependencies {
     implementation (libs.blankj.utilcodex)
     implementation (projects.common.preference)
     implementation (projects.common.resources)
-    implementation (projects.featureSettings)
     implementation (projects.domain)
     implementation (projects.data)
     implementation(libs.androidx.material)
