+++ feature_rpc_base/src/main/java/com/my/kizzy/feature_rpc_base/services/AppDetectionService.kt
@@ -26,9 +26,9 @@ import com.google.gson.reflect.TypeToken
 import com.my.kizzy.data.rpc.Constants
 import com.my.kizzy.data.rpc.KizzyRPC
 import com.my.kizzy.data.rpc.RpcImage
-import com.my.kizzy.resources.R
-import com.my.kizzy.feature_settings.rpc_settings.RpcButtons
+import com.my.kizzy.domain.model.RpcButtons
 import com.my.kizzy.preference.Prefs
+import com.my.kizzy.resources.R
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.*
 import java.util.*
